+++ src/blob/handlers/ServiceHandler.ts
@@ -9,9 +9,16 @@ import {
   BLOB_API_VERSION,
   DEFAULT_LIST_CONTAINERS_MAX_RESULTS,
   EMULATOR_ACCOUNT_KIND,
-  EMULATOR_ACCOUNT_SKUNAME
+  EMULATOR_ACCOUNT_SKUNAME,
 } from "../utils/constants";
 import BaseHandler from "./BaseHandler";
+import IAccountDataStore from "../../common/IAccountDataStore";
+import IExtentStore from "../../common/persistence/IExtentStore";
+import IBlobMetadataStore from "../persistence/IBlobMetadataStore";
+import ILogger from "../../common/ILogger";
+import { BlobBatchHandler } from "./BlobBatchHandler";
+import { Readable } from "stream";
+import { OAuthLevel } from "../../common/models";
 
 /**
  * ServiceHandler handles Azure Storage Blob service related requests.
@@ -22,6 +29,17 @@ import BaseHandler from "./BaseHandler";
  */
 export default class ServiceHandler extends BaseHandler
   implements IServiceHandler {
+
+  constructor(
+      private readonly accountDataStore: IAccountDataStore,
+      private readonly oauth: OAuthLevel | undefined,
+      metadataStore: IBlobMetadataStore,
+      extentStore: IExtentStore,
+      logger: ILogger,
+      loose: boolean
+    ) {
+      super(metadataStore, extentStore, logger, loose);
+    }
   /**
    * Default service properties.
    *
@@ -67,14 +85,41 @@ export default class ServiceHandler extends BaseHandler
     throw new NotImplementedError(context.contextId);
   }
 
-  public submitBatch(
+
+  public async submitBatch(
     body: NodeJS.ReadableStream,
     contentLength: number,
     multipartContentType: string,
     options: Models.ServiceSubmitBatchOptionalParams,
     context: Context
   ): Promise<Models.ServiceSubmitBatchResponse> {
-    throw new NotImplementedError(context.contextId);
+    const blobServiceCtx = new BlobStorageContext(context);
+    const requestBatchBoundary = blobServiceCtx.request!.getHeader("content-type")!.split("=")[1];
+
+    const blobBatchHandler = new BlobBatchHandler(this.accountDataStore, this.oauth,
+       this.metadataStore, this.extentStore, this.logger, this.loose);
+
+    const responseBodyString = await blobBatchHandler.submitBatch(body,
+      requestBatchBoundary,
+      "",
+      context.request!,
+      context);
+
+    const responseBody = new Readable();
+    responseBody.push(responseBodyString);
+    responseBody.push(null);
+
+    // No client request id defined in batch response, should refine swagger and regenerate from it.
+    // batch response succeed code should be 202 instead of 200, should refine swagger and regenerate from it.
+    const response: Models.ServiceSubmitBatchResponse = {
+      statusCode: 202,
+      requestId: context.contextId,
+      version: BLOB_API_VERSION,
+      contentType: "multipart/mixed; boundary=" + requestBatchBoundary,
+      body: responseBody
+    };
+
+    return response;
   }
 
   /**