+++ src/blob/utils/constants.ts
@@ -1,8 +1,8 @@
 import { StoreDestinationArray } from "../../common/persistence/IExtentStore";
 import * as Models from "../generated/artifacts/models";
 
-export const VERSION = "3.17.1";
-export const BLOB_API_VERSION = "2021-06-08";
+export const VERSION = "3.20.1";
+export const BLOB_API_VERSION = "2021-10-04";
 export const DEFAULT_BLOB_SERVER_HOST_NAME = "127.0.0.1"; // Change to 0.0.0.0 when needs external access
 export const DEFAULT_LIST_BLOBS_MAX_RESULTS = 5000;
 export const DEFAULT_LIST_CONTAINERS_MAX_RESULTS = 5000;
@@ -96,6 +96,8 @@ export const DEFAULT_BLOB_PERSISTENCE_ARRAY: StoreDestinationArray = [
 ];
 
 export const ValidAPIVersions = [
+  "2021-10-04",
+  "2021-08-06",
   "2021-06-08",
   "2021-04-10",
   "2021-02-12",
@@ -151,3 +153,6 @@ export const VALID_BLOB_AUDIENCES = [
   /^https:\/\/(.*)\.blob\.core\.usgovcloudapi\.net[\/]?$/,
   /^https:\/\/(.*)\.blob\.core\.cloudapi\.de[\/]?$/
 ];
+
+export const HTTP_LINE_ENDING = "\r\n";
+export const HTTP_HEADER_DELIMITER = ": ";