+++ src/table/errors/StorageErrorFactory.ts
@@ -27,6 +27,20 @@ export default class StorageErrorFactory {
     );
   }
 
+  public static getInvalidAPIVersion(
+    context: Context,
+    apiVersion?: string,
+  ): StorageError {
+    return new StorageError(
+      400,
+      "InvalidHeaderValue",
+      `The API version ${apiVersion} is not supported by Azurite. Please upgrade Azurite to latest version and retry. If you are using Azurite in Visual Studio, please check you have installed latest Visual Studio patch. Azurite command line parameter \"--skipApiVersionCheck\" or Visual Studio Code configuration \"Skip Api Version Check\" can skip this error. `,
+      context.contextID || defaultID,
+      undefined,
+      context
+    );
+  }
+
   public static getInvalidInput(
     context: Context,
     additionalMessages?: { [key: string]: string }