+++ src/table/utils/constants.ts
@@ -17,8 +17,8 @@ export enum TABLE_STATUSCODE {
 }
 
 export const DEFAULT_TABLE_CONTEXT_PATH = "azurite_table_context";
-export const TABLE_API_VERSION = "2021-06-08";
-export const VERSION = "3.17.1";
+export const TABLE_API_VERSION = "2021-10-04";
+export const VERSION = "3.20.1";
 // Max Body size is 4 MB
 export const BODY_SIZE_MAX = 1024 * 1024 * 4;
 
@@ -71,6 +71,8 @@ export const DEFAULT_TABLE_PERSISTENCE_ARRAY: StoreDestinationArray = [
 
 export const QUERY_RESULT_MAX_NUM = 1000;
 export const ValidAPIVersions = [
+  "2021-10-04",
+  "2021-08-06",
   "2021-06-08",
   "2021-04-10",
   "2021-02-12",