+++ tests/table/apis/table.entity.test.ts
@@ -336,10 +336,10 @@ describe("table Entity APIs test - using Azure-Storage", () => {
       tableName,
       entityToUpdate,
       (updateError, updateResult, updateResponse) => {
-        if (updateError) {
-          assert.fail("Test threw an error : " + updateError);
+        if (!updateError) {
+          assert.fail("Test should have thrown an error");
         } else {
-          assert.strictEqual(updateResponse.statusCode, 204);
+          assert.strictEqual(updateResponse.statusCode, 404);
         }
         done();
       }
@@ -604,18 +604,22 @@ describe("table Entity APIs test - using Azure-Storage", () => {
   });
 
   [
+    { pk: "pk", rk: "rk", label: "normal partition key and row key" },
     { pk: "", rk: "rk", label: "empty partition key" },
-    { pk: "pk", rk: "", label: "empty row key" },
+    { pk: "pk", rk: "", label: "empty row key" }
   ].forEach(({ pk, rk, label }) => {
-
-    ["INSERT", "INSERT_OR_MERGE", "INSERT_OR_REPLACE"].forEach(operation => {
+    ["INSERT", "INSERT_OR_MERGE", "INSERT_OR_REPLACE"].forEach((operation) => {
       it(`${operation} entity with ${label} in a BATCH, @loki`, (done) => {
         requestOverride.headers = {
           Prefer: "return-content",
           accept: "application/json;odata=fullmetadata"
         };
 
-        const batchEntity1 = new TestEntity(!pk ? pk : getUniqueName(pk), !rk ? rk : getUniqueName(rk), "value1");
+        const batchEntity1 = new TestEntity(
+          !pk ? pk : getUniqueName(pk),
+          !rk ? rk : getUniqueName(rk),
+          "value1"
+        );
 
         const entityBatch: Azure.TableBatch = new Azure.TableBatch();
         entityBatch.addOperation(operation, batchEntity1);
@@ -638,7 +642,10 @@ describe("table Entity APIs test - using Azure-Storage", () => {
                     assert.ifError(error);
                   } else if (result) {
                     const entity: TestEntity = result;
-                    assert.strictEqual(entity.myValue._, batchEntity1.myValue._);
+                    assert.strictEqual(
+                      entity.myValue._,
+                      batchEntity1.myValue._
+                    );
                   }
                   done();
                 }
@@ -649,14 +656,49 @@ describe("table Entity APIs test - using Azure-Storage", () => {
       });
     });
 
-    ["MERGE", "REPLACE"].forEach(operation => {
+    ["MERGE", "REPLACE"].forEach((operation) => {
+      it(`${operation} of non-existent entity with ${label} in a BATCH, @loki`, (done) => {
+        requestOverride.headers = {
+          Prefer: "return-content",
+          accept: "application/json;odata=fullmetadata"
+        };
+
+        const batchEntity1 = new TestEntity(
+          !pk ? pk : getUniqueName(pk),
+          !rk ? rk : getUniqueName(rk),
+          "value1"
+        );
+
+        const entityBatch: Azure.TableBatch = new Azure.TableBatch();
+        entityBatch.addOperation(operation, batchEntity1);
+
+        tableService.executeBatch(
+          tableName,
+          entityBatch,
+          (updateError, updateResult, updateResponse) => {
+            if (!updateError) {
+              assert.fail("Test should have thrown an error");
+            } else {
+              assert.strictEqual(updateResponse.statusCode, 404);
+              done();
+            }
+          }
+        );
+      });
+    });
+
+    ["MERGE", "REPLACE"].forEach((operation) => {
       it(`${operation} of entity with ${label} in a BATCH, @loki`, (done) => {
         requestOverride.headers = {
           Prefer: "return-content",
           accept: "application/json;odata=fullmetadata"
         };
 
-        const batchEntity1 = new TestEntity(!pk ? pk : getUniqueName(pk), !rk ? rk : getUniqueName(rk), "value1");
+        const batchEntity1 = new TestEntity(
+          !pk ? pk : getUniqueName(pk),
+          !rk ? rk : getUniqueName(rk),
+          "value1"
+        );
 
         tableService.insertEntity(
           tableName,
@@ -680,7 +722,8 @@ describe("table Entity APIs test - using Azure-Storage", () => {
                 }
               }
             );
-          })
+          }
+        );
       });
     });
 
@@ -690,7 +733,11 @@ describe("table Entity APIs test - using Azure-Storage", () => {
         accept: "application/json;odata=fullmetadata"
       };
 
-      const batchEntity1 = new TestEntity(!pk ? pk : getUniqueName(pk), !rk ? rk : getUniqueName(rk), "value1");
+      const batchEntity1 = new TestEntity(
+        !pk ? pk : getUniqueName(pk),
+        !rk ? rk : getUniqueName(rk),
+        "value1"
+      );
 
       tableService.insertEntity(
         tableName,
@@ -722,7 +769,8 @@ describe("table Entity APIs test - using Azure-Storage", () => {
               }
             }
           );
-        })
+        }
+      );
     });
   });