+++ tests/table/apis/table.validation.rest.test.ts
@@ -9,7 +9,11 @@ import { configLogger } from "../../../src/common/Logger";
 import TableConfiguration from "../../../src/table/TableConfiguration";
 import TableServer from "../../../src/table/TableServer";
 import { getUniqueName } from "../../testutils";
-import { postToAzurite } from "../utils/table.entity.tests.rest.submitter";
+import {
+  deleteToAzurite,
+  getToAzurite,
+  postToAzurite
+} from "../utils/table.entity.tests.rest.submitter";
 
 // Set true to enable debug log
 configLogger(false);
@@ -17,7 +21,7 @@ configLogger(false);
 describe("table name validation tests", () => {
   const host = "127.0.0.1";
   const port = 11002;
-  const metadataDbPath = "__tableTestsStorage__";
+  const metadataDbPath = getUniqueName("__tableTestsStorage__");
   const enableDebugLog: boolean = true;
   const debugLogPath: string = "g:/debug.log";
   const config = new TableConfiguration(
@@ -226,6 +230,166 @@ describe("table name validation tests", () => {
     });
     try {
       await postToAzurite("Tables", body3, createTableHeaders);
+    } catch (err: any) {
+      assert.strictEqual(
+        err,
+        undefined,
+        `unexpected exception with start trimmed! : ${err}`
+      );
+    }
+  });
+
+  it("should delete a table with a name which is a substring of an existing table, @loki", async () => {
+    tableName = getUniqueName("table");
+    const shortName = tableName.substring(0, 6);
+    const basicBody = JSON.stringify({
+      TableName: shortName
+    });
+    const createTableHeaders = {
+      "Content-Type": "application/json",
+      Accept: "application/json;odata=nometadata"
+    };
+
+    // get the table count before test
+    const tableCountBeforeTest = (await getToAzurite("Tables", createTableHeaders)).data.value.length;
+
+    // create table 1
+    try {
+      const createTable1 = await postToAzurite(
+        "Tables",
+        basicBody,
+        createTableHeaders
+      );
+      assert.strictEqual(
+        createTable1.status,
+        201,
+        "Did not have the expected number of tables before deletion"
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err.response.status,
+        201,
+        `unexpected status code creating first table : ${err.response.status}`
+      );
+    }
+    const body = JSON.stringify({
+      TableName: tableName
+    });
+    // create table 2
+    try {
+      const createTable2 = await postToAzurite(
+        "Tables",
+        body,
+        createTableHeaders
+      );
+      assert.strictEqual(
+        createTable2.status,
+        201,
+        "Did not have the expected number of tables before deletion"
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err.response.status,
+        201,
+        `unexpected status code : ${err.response.status}`
+      );
+    }
+    const tableName2 = tableName.substring(0, tableName.length - 4);
+    const body2 = JSON.stringify({
+      TableName: tableName2
+    });
+    // create table 3
+    try {
+      const createTable3 = await postToAzurite(
+        "Tables",
+        body2,
+        createTableHeaders
+      );
+      assert.strictEqual(
+        createTable3.status,
+        201,
+        "Did not have the expected number of tables before deletion"
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err,
+        undefined,
+        `unexpected exception with end trimmed! : ${err}`
+      );
+    }
+    const tableName3 = tableName.substring(4);
+    const body3 = JSON.stringify({
+      TableName: tableName3
+    });
+    // create table 4
+    try {
+      const createTable4 = await postToAzurite(
+        "Tables",
+        body3,
+        createTableHeaders
+      );
+      assert.strictEqual(
+        createTable4.status,
+        201,
+        "Did not have the expected number of tables before deletion"
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err,
+        undefined,
+        `unexpected exception with start trimmed! : ${err}`
+      );
+    }
+    // now list tables after deletion...
+    try {
+      const listTableResult1 = await getToAzurite("Tables", createTableHeaders);
+      // we count all tables created in the tests before this one as well
+      assert.strictEqual(
+        listTableResult1.data.value.length,
+        tableCountBeforeTest + 4,
+        "Did not have the expected number of tables before deletion"
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err,
+        undefined,
+        `unexpected exception with start trimmed! : ${err}`
+      );
+    }
+    // now delete "table"
+    try {
+      const deleteResult = await deleteToAzurite(
+        `Tables('${shortName}')`,
+        "",
+        createTableHeaders
+      );
+      assert.strictEqual(
+        deleteResult.status,
+        204,
+        "Delete was not successful."
+      );
+    } catch (err: any) {
+      assert.strictEqual(
+        err,
+        undefined,
+        `unexpected exception with start trimmed! : ${err}`
+      );
+    }
+    // now list tables after deletion...
+    try {
+      const listTableResult2 = await getToAzurite("Tables", createTableHeaders);
+      assert.strictEqual(
+        listTableResult2.data.value.length,
+        tableCountBeforeTest + 3,
+        "Did not have the expected number of tables after deletion"
+      );
+      for (const table of listTableResult2.data.value) {
+        assert.notStrictEqual(
+          table.TableName,
+          shortName,
+          "We still list the table we should have deleted."
+        );
+      }
     } catch (err: any) {
       assert.strictEqual(
         err,