+++ ChangeLog.md
@@ -4,12 +4,61 @@
 
 ## Upcoming Release
 
+## 2022.10 Version 3.20.1
+
+General:
+
+- Bump package version.
+
+## 2022.10 Version 3.20.0
+
+General:
+
+- Make emulator start commands async so that they can be awaited by clients.
+
+Blob:
+
+- Add support for blob batch operation.
+
+Table:
+
+- TimeStamp and Etag use the same high precision value as source.
+
+## 2022.09 Version 3.19.0
+
+General:
+
+- Bump up service API version to 2021-10-04
+- Added support for docker image on arm64 architecture.
+- Updated Readme by adding account key must be base64 encoded string.
+
+Table:
+
+- Correctly responds with status 202 on merge with non-existent entity.
+- Properly differentiate between upsert and update in batch merge and replace.
+- Added additional tests via raw REST tests.
+- Correctly deletes a table that is a substring of another table.
+- Adds Sample Go App to check that Batch responses work for Go SDK.
+- Removes extra CRLFs from all serialized Batch responses, adds missing CRLF after Etag header.
+
+## 2022.06 Version 3.18.0
+
+General:
+
+- Bump up service API version to 2021-08-06
+- Modified the error messge for invalid API version to make it more actionable.
+
+Blob:
+
+- Fixed issue that startCopyFromURL and copyFromURL API not fail, when request container if-none-match="\*" and dest blob already exist.
+
 Table:
 
 - Reject table batch request bodies exceeding 4MB.
 - Fix binary table property validation to be 64K bytes not 32K characters.
 - Does not error when table created is a substring of another table.
 - Correctly responds with status 404 on patch with non-existant entity.
+- Fix pagination when no rowkey in continuation token
 
 ## 2022.04 Version 3.17.1